Convert regexps' capturing groups to non-capturing groups
Normally within a regexp '(' does not refer to a char itself, but starts
"a capturing group". (String within a capturing group gets stored somewhere.)
Lint does not like this for some reason. Therefore, if you are specifically
interested to get the content within parentheses, make change the capturing
group to a non-capturing group by adding '?:' after '('.

Thus /(foo)/ becomes /(?:foo)/. Note that ? does not get confused with it's
regexp meaning "zero-or-none of the preceding stuff", since preceding stuff is "open catching/non-catching group" character, and not normal text, a range or
a group.

Minuses: This does not improve code readability (just the opposite), so I
personally think it's a stupid error.
Plusses: Apparently adding "?:" makes the regexp easier and faster to compile,
and possible requires less memory.